+++ nextjs/src/app/[locale]/(pages)/[...slug]/page.tsx
@@ -17,11 +17,28 @@ export async function generateMetadata({
     slug: string[]
   }
 }): Promise<Metadata> {
-  console.log({ locale, slug })
-  const data = await getPage(locale, slug.join("/"))
+  const path = slug.join("/")
+  const data = await getPage(locale, path)
+  const languages = data.localizations.reduce(
+    (acc: any, item: any) => {
+      const slugLocale = item.locale.toLowerCase()
+      acc[item.locale] = `/${slugLocale}/${item.slug}`
+      return acc
+    },
+    { [locale]: `/${locale}/${slug}` },
+  )
+
+  if (languages?.en !== undefined) {
+    languages["x-default"] = languages.en
+  }
+
   return {
     title: data.metadata_title,
     description: data.metadata_description,
+    alternates: {
+      canonical: `/${locale}/${path}`,
+      languages,
+    },
   }
 }